Add files via upload
Dr. Claudia Roessing presented the VISION Project at the Digital Academy for the Sustainable Built Environment (DASBE), Transforming Construction Skills Conference in Portlaoise, Ireland. The project is funded by EPA and its objective is to investigate and conduct a feasibility study to assess the use, application and effectiveness of economic and regulatory instruments to reduce the use of virgin materials in the construction sector in Ireland.